# mj_web_openapi.yaml
openapi: 3.0.3
info:
  title: MJ_WEB API
  version: 1.0.0
servers:
- url: /api
paths:
...
  /products:
    get:
      summary: List products
      responses:
        '200':
          description: OK
      parameters:
      - in: query
        name: q
        schema:
          type: string
  /products/{id}:
    get:
      summary: Get product detail
      responses:
        '200':
          description: OK
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
  /products/{id}/related:
    get:
      summary: Related products
      responses:
        '200':
          description: OK
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT